tui: move what every screen shares into tui/context.py
`settings.py`, `overview.py`, `app.py` and `cli.py` each reached into a screen
module for `Context` and the footer, so the shared types now have a module of
their own that imports neither. `_say` becomes `say`: two callers already
imported it across modules.
