+++ tests/unit/test_every_row.py
@@ -19,7 +19,7 @@ import pytest
 
 from gentoo_install.model.config import InstallConfig
 from gentoo_install.tui import app, settings
-from gentoo_install.tui.screens import Context
+from gentoo_install.tui.context import Context
 from gentoo_install.tui.settings import UNSET, Setting
 from gentoo_install.tui.widgets import Answer, Outcome
